requirements.txt
biopython==1.79
scikit-learn==1.3.0
tensorflow==2.15.0
